refactor(data, ui): remove old database stuffs
Also replace the new database stuffs.
