+++ app/src/main/AndroidManifest.xml
@@ -65,10 +65,6 @@
             android:launchMode="singleTask"
             android:parentActivityName=".ui.main.PrefMainActivity" />
 
-        <activity
-            android:name=".ui.main.LiquidKeyboardActivity"
-            android:parentActivityName=".ui.main.PrefMainActivity" />
-
         <!-- Using an activity alias to disable/enable the app icon in the launcher -->
         <activity-alias
             android:name="com.osfans.trime.PrefLauncherAlias"
