+++ app/src/main/java/com/osfans/trime/TrimeApplication.kt
@@ -5,9 +5,9 @@ import android.os.Process
 import androidx.preference.PreferenceManager
 import cat.ereza.customactivityoncrash.config.CaocConfig
 import com.osfans.trime.data.AppPrefs
+import com.osfans.trime.data.db.ClipboardHelper
 import com.osfans.trime.data.db.CollectionHelper
-import com.osfans.trime.data.db.clipboard.ClipboardHelper
-import com.osfans.trime.data.db.draft.DraftHelper
+import com.osfans.trime.data.db.DraftHelper
 import com.osfans.trime.ui.main.LogActivity
 import timber.log.Timber
