+++ app/src/main/java/com/osfans/trime/data/db/DraftHelper.kt
@@ -1,11 +1,8 @@
-package com.osfans.trime.data.db.draft
+package com.osfans.trime.data.db
 
 import android.content.Context
 import androidx.room.Room
 import com.osfans.trime.data.AppPrefs
-import com.osfans.trime.data.db.Database
-import com.osfans.trime.data.db.DatabaseBean
-import com.osfans.trime.data.db.DatabaseDao
 import com.osfans.trime.ime.core.Trime
 import com.osfans.trime.util.StringUtils.mismatch
 import kotlinx.coroutines.CoroutineScope
