--- app/src/main/java/com/osfans/trime/data/db/clipboard/ClipboardDao.java
@@ -1,96 +1,0 @@
-package com.osfans.trime.data.db.clipboard;
-
-import android.database.Cursor;
-import android.database.sqlite.SQLiteDatabase;
-import androidx.annotation.NonNull;
-import com.osfans.trime.data.db.DbBean;
-import com.osfans.trime.data.db.DbHelper;
-import com.osfans.trime.ime.core.Trime;
-import com.osfans.trime.ime.symbol.SimpleKeyBean;
-import java.util.ArrayList;
-import java.util.List;
-import timber.log.Timber;
-
-public class ClipboardDao {
-
-  private DbHelper helper;
-  private static ClipboardDao self;
-  private static String DB_NAME = "clipboard.db";
-
-  public static ClipboardDao get() {
-    if (null == self) self = new ClipboardDao();
-    return self;
-  }
-
-  public ClipboardDao() {}
-
-  /** 插入新记录 * */
-  public void insert(@NonNull DbBean clipboardBean) {
-    helper = new DbHelper(Trime.getService(), DB_NAME);
-    SQLiteDatabase db = helper.getWritableDatabase();
-    db.execSQL(
-        "insert into t_data(text,html,type,time) values(?,?,?,?)",
-        new Object[] {
-          clipboardBean.getText(),
-          clipboardBean.getHtml(),
-          clipboardBean.getType(),
-          clipboardBean.getTime()
-        });
-    db.close();
-  }
-
-  /** 删除文字相同的剪贴板记录，插入新记录 * */
-  public void add(@NonNull DbBean clipboardBean) {
-    helper = new DbHelper(Trime.getService(), DB_NAME);
-    SQLiteDatabase db = helper.getWritableDatabase();
-    db.delete("t_data", "text=?", new String[] {clipboardBean.getText()});
-    db.execSQL(
-        "insert into t_data(text,html,type,time) values(?,?,?,?)",
-        new Object[] {
-          clipboardBean.getText(),
-          clipboardBean.getHtml(),
-          clipboardBean.getType(),
-          clipboardBean.getTime()
-        });
-    db.close();
-  }
-
-  /** 删除记录 * */
-  public void delete(@NonNull String str) {
-    helper = new DbHelper(Trime.getService(), DB_NAME);
-    SQLiteDatabase db = helper.getWritableDatabase();
-    db.delete("t_data", "text=?", new String[] {str});
-    db.close();
-  }
-
-  public void delete(@NonNull List<SimpleKeyBean> list) {
-    helper = new DbHelper(Trime.getService(), DB_NAME);
-    SQLiteDatabase db = helper.getWritableDatabase();
-    for (SimpleKeyBean bean : list) db.delete("t_data", "text=?", new String[] {bean.getText()});
-    db.close();
-  }
-
-  public List<SimpleKeyBean> getAllSimpleBean(int size) {
-
-    List<SimpleKeyBean> list = new ArrayList<>();
-    if (size == 0) return list;
-
-    String sql = "select text , html , type , time from t_data ORDER BY time DESC";
-    if (size > 0) sql = sql + " limit 0," + size;
-
-    helper = new DbHelper(Trime.getService(), DB_NAME);
-
-    SQLiteDatabase db = helper.getWritableDatabase();
-    Cursor cursor = db.rawQuery(sql, null);
-    if (cursor != null) {
-      while (cursor.moveToNext()) {
-        DbBean v = new DbBean(cursor.getString(0));
-        list.add(v);
-      }
-      cursor.close();
-    }
-    db.close();
-    Timber.d("getAllSimpleBean() size=%s limit=%s", list.size(), size);
-    return list;
-  }
-}
