+++ app/src/main/res/xml/other_preference.xml
@@ -38,10 +38,6 @@
             app:title="@string/other_clipboard_limit"
             app:useSimpleSummaryProvider="true" />
 
-        <Preference android:key="other_managed_clipboard"
-            app:iconSpaceReserved="false"
-            android:title="@string/other_managed_clipboard"/>
-
         <EditTextPreference android:key="other__clipboard_compare"
             app:iconSpaceReserved="false"
             android:title="@string/other__clipboard_compare_title"
@@ -52,14 +48,6 @@
             android:title="@string/other__clipboard_output_title"
             android:defaultValue=""
             app:useSimpleSummaryProvider="true"/>
-    </PreferenceCategory>
-
-    <PreferenceCategory android:key="other_collection"
-        app:iconSpaceReserved="false"
-        app:title="@string/other_collection">
-        <Preference android:key="other_managed_collection"
-            app:iconSpaceReserved="false"
-            android:title="@string/other_managed_collection"/>
     </PreferenceCategory>
 
     <PreferenceCategory android:key="other_draft"
@@ -74,10 +62,6 @@
             app:systemDefaultValueText="@string/disable"
             app:title="@string/other_draft_limit"
             app:useSimpleSummaryProvider="true" />
-
-        <Preference android:key="other_managed_draft"
-            app:iconSpaceReserved="false"
-            android:title="@string/other_managed_draft"/>
 
         <EditTextPreference android:key="other__draft_output"
             app:iconSpaceReserved="false"
